Setupping up python project configuration
Finally, the python project is a package where it can be installed, so
the imports in everywhere else in the code doesn't broke while try to
running.
