+++ python/requirements.txt
@@ -1,9 +1,7 @@
-autopep8==1.6.0
-mpmath==1.2.1
+jupyter-core==4.11.1
+jupyter_client==7.3.5
+matplotlib==3.5.3
+matplotlib-inline==0.1.6
 numpy==1.23.2
-pycodestyle==2.8.0
-snakeviz==2.1.1
 symengine==0.9.2
 sympy==1.10.1
-toml==0.10.2
-tornado==6.2
