+++ python/setup.cfg
@@ -1,0 +1,29 @@
+[metadata]
+name = rman
+description = Robotic Manipulation revamp project in python from the course in the second semester of 2021/22
+author = filipe-varela
+license = Apache License
+license_file = ../LICENSE
+platforms = unix, linux, osx, cygwin, win32
+classifiers = 
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3 :: Only
+    Programming Language :: Python :: 3 :: 3.7
+    Programming Language :: Python :: 3 :: 3.8
+    Programming Language :: Python :: 3 :: 3.9
+
+[options]
+packages = 
+    rman
+install_requires =
+    sympy>=1.11
+    symengine>=0.9
+    numpy>=1.23
+    matplotlib>=3.5
+    jupyter-core>=4.11
+    jupyter_client>=7.3
+python_requires = >=3.7
+package_dir = 
+    =src
+zip_safe = no
+
