--- python/tests/__init__.py
@@ -1,1 +1,0 @@
-# from .test_kinematic import TestKinematic
